services finalizados, todos se comunicando entre si. templates e controllers iniciais já cadastrando empresa, oferta de estágio, listando ofertas e empresas.

+++ src/main/java/com/eqp3e1/repository/CoordenadorRepository.java
@@ -3,5 +3,5 @@ package com.eqp3e1.repository;
 import com.eqp3e1.model.Coordenador;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CoordenadorRepository extends CrudRepository<Coordenador, Integer> {
+public interface CoordenadorRepository extends CrudRepository<Coordenador, Long> {
 }
