+++ src/main/java/com/eqp3e1/repository/EstagioRepository.java
@@ -3,5 +3,5 @@ package com.eqp3e1.repository;
 import com.eqp3e1.model.Estagio;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EstagioRepository extends CrudRepository<Estagio, Integer> {
+public interface EstagioRepository extends CrudRepository<Estagio, Long> {
 }
