+++ src/main/java/com/eqp3e1/repository/HabilidadeRepository.java
@@ -3,5 +3,5 @@ package com.eqp3e1.repository;
 import com.eqp3e1.model.Habilidade;
 import org.springframework.data.repository.CrudRepository;
 
-public interface HabilidadeRepository extends CrudRepository<Habilidade, Integer> {
+public interface HabilidadeRepository extends CrudRepository<Habilidade, Long> {
 }
